fix: image delete logic

+++ src/main/java/com/qnnect/user/service/UserServiceImpl.java
@@ -67,7 +67,7 @@ public class UserServiceImpl implements UserService{
     }
 
     public void updateToDefaultImage(User user){
-        if(user.getProfilePicture() != null || user.getProfilePicture() != defaultImage){
+        if(user.getProfilePicture() != null && user.getProfilePicture() != defaultImage){
             s3Uploader.deleteS3(user.getProfilePicture(), "profile");
         }
         user.setProfilePicture(defaultImage);
